various optimizations, code cleanups
- separate helpers
- batching
- cleaner logs
